评论1

+++ resources/views/post/show.blade.php
@@ -5,15 +5,19 @@
         <div class="blog-post">
             <div style="display:inline-flex">
                 <h2 class="blog-post-title"><?php echo $post->title; ?></h2>
+                @can('update', $post)
                     <a style="margin: auto" href="/posts/{{$post->id}}/edit">
                         <span class="glyphicon glyphicon-pencil" aria-hidden="true"></span>
                     </a>
-                <a style="margin: auto" href="/posts/{{$post->id}}/delete">
-                    <span class="glyphicon glyphicon-remove" aria-hidden="true"></span>
-                </a>
+                @endcan
+                @can('delete', $post)
+                    <a style="margin: auto" href="/posts/{{$post->id}}/delete">
+                        <span class="glyphicon glyphicon-remove" aria-hidden="true"></span>
+                    </a>
+                @endcan
             </div>
 
-            <p class="blog-post-meta">{{$post->created_at->toFormattedDateString()}}<a href="#">Kassandra Ankunding2</a></p>
+            <p class="blog-post-meta">{{$post->created_at->toFormattedDateString()}}<a href="#"> <?php echo $post->user->name ?? 'xxx'; ?></a></p>
 
             <p>
             <p>{!! $post->content !!}</p>
@@ -29,12 +33,14 @@
 
             <!-- List group -->
             <ul class="list-group">
-                <li class="list-group-item">
-                    <h5>2017-05-28 10:15:08 by Kassandra Ankunding2</h5>
-                    <div>
-                        这是第一个评论这是第一个评论这是第一个评论这是第一个评论这是第一个评论这是第一个评论这是第一个评论这是第一个评论这是第一个评论
-                    </div>
-                </li>
+                @foreach($post->comments as $comment)
+                    <li class="list-group-item">
+                        <h5>{{$comment->created_at}} by </h5>
+                        <div>
+                            {{$comment->content}}
+                        </div>
+                    </li>
+                @endforeach
             </ul>
         </div>
 
@@ -45,12 +51,23 @@
             <!-- List group -->
             <ul class="list-group">
                 <form action="/posts/comment" method="post">
-                    <input type="hidden" name="_token" value="4BfTBDF90Mjp8hdoie6QGDPJF2J5AgmpsC9ddFHD">
-                    <input type="hidden" name="post_id" value="62"/>
+                    {{csrf_field()}}
+                    <input type="hidden" name="post_id" value="{{$post->id}}"/>
                     <li class="list-group-item">
                         <textarea name="content" class="form-control" rows="10"></textarea>
+                        @if(count($errors)>0)
+                            <div class="alert alert-danger" role="alert">
+                                @foreach($errors->all() as $error)
+                                    <p>{{$error}}</p>
+                                @endforeach
+                            </div>
+                        @endif
                         <button class="btn btn-default" type="submit">提交</button>
+
                     </li>
+                    {{-- @include("layout.error")--}}
+
+
                 </form>
 
             </ul>
